Gradient_Descent_SV- Description
This project synthesizes a number of samples before creating a model to separate the samples into two groups, training the model using alpha descent. The program notifies the user after the program has completed a full round of looking at the samples and eventually stops after the change from one iteration to another becomes small enough that the time used is not requited by the higher accuracy that the model would acheive.