Stochastic_Descent_SV- Description
This program takes in a number of epochs to run, a batch size, and an alpha descent value to apply Stochastic Gradient Descent on a generated dataset made up of 1000 sample points. The program applies Stochastic Gradient Descent to achieve an extremely high accuracy and displays the final results of the model training to the console.